Documents/GitHub/client_dashboard/requirements.txt
openai
airtable-python-wrapper
pandas
pyyaml
beautifulsoup4
requests
dash==2.0.0
dash-bootstrap-components==1.0.0
plotly==5.4.0
scikit-learn==1.0
werkzeug==2.0.0
numpy>=1.24.0